refactor(mm): hoist shared embed_input_ids into RBLNOptimumMultimodalMixin
The six simple MM models shared the same scatter-merge logic; they only
differed on (1) the placeholder token attr name and (2) whether OOV
tokens need PAD masking before the text embedding lookup. Move the merge
into the mixin, parameterized by two small hooks:

  - _image_token_id(): defaults to config.image_token_index
  - _embed_text_tokens(): defaults to a plain in-vocab lookup

llava/llava_next/blip2 now use the defaults outright; idefics3 overrides
only the token-id hook; paligemma/gemma3 override _embed_text_tokens to
keep their exact OOV PAD-masking (preserving each model's own vocab-size
source). Net -98 lines, behavior unchanged.
