+++ vllm_rbln/model_executor/models/optimum/llava.py
@@ -16,7 +16,6 @@ from typing import Union
 import torch
 from vllm.config import VllmConfig
 from vllm.logger import init_logger
-from vllm.model_executor.models.interfaces import MultiModalEmbeddings
 from vllm.model_executor.models.llava import (
     LlavaImageInputs,
     LlavaImagePixelInputs,
@@ -154,31 +153,6 @@ class RBLNOptimumLlavaForConditionalGeneration(
             image_sizes=image_sizes,
         )
         return list(image_features)
-
-    def embed_input_ids(
-        self,
-        input_ids: torch.Tensor,
-        multimodal_embeddings: MultiModalEmbeddings | None = None,
-        *,
-        is_multimodal: torch.Tensor | None = None,
-    ) -> torch.Tensor:
-        # Mirrors optimum-rbln's _preprocess_prefill: image tokens are in-vocab
-        # for LLaVA, so no PAD masking is needed; just scatter the image
-        # features over the image-token positions.
-        config = self.model.config
-        if is_multimodal is None:
-            is_multimodal = input_ids == config.image_token_index
-
-        inputs_embeds = self.model.get_input_embeddings()(input_ids)
-
-        if multimodal_embeddings is None or len(multimodal_embeddings) == 0:
-            return inputs_embeds
-
-        mm_embeds = torch.cat(list(multimodal_embeddings)).to(
-            inputs_embeds.device, inputs_embeds.dtype
-        )
-        scatter_mask = is_multimodal.unsqueeze(-1).expand_as(inputs_embeds)
-        return inputs_embeds.masked_scatter(scatter_mask, mm_embeds)
 
     def _parse_and_validate_image_input(
         self, **kwargs: object
