+++ vllm_rbln/model_executor/models/optimum/paligemma.py
@@ -15,7 +15,6 @@ from typing import Any
 
 import torch
 from vllm.config import VllmConfig
-from vllm.model_executor.models.interfaces import MultiModalEmbeddings
 from vllm.model_executor.models.paligemma import (
     PaliGemmaImageEmbeddingInputs,
     PaliGemmaImageInputs,
@@ -124,35 +123,19 @@ class RBLNOptimumPaliGemmaForConditionalGeneration(
         image_features = self.model.get_image_features(image_input["data"])
         return list(image_features)
 
-    def embed_input_ids(
-        self,
-        input_ids: torch.Tensor,
-        multimodal_embeddings: MultiModalEmbeddings | None = None,
-        *,
-        is_multimodal: torch.Tensor | None = None,
+    def _image_token_id(self) -> int:
+        # PaliGemma's HF config names the placeholder `image_token_id`.
+        return self.model.config.image_token_id
+
+    def _embed_text_tokens(
+        self, input_ids: torch.Tensor, is_multimodal: torch.Tensor
     ) -> torch.Tensor:
-        # Mirrors the merge semantics of optimum-rbln's _preprocess_prefill:
-        # replace OOV image tokens with PAD for the text embedding lookup,
-        # then scatter the image features over the image-token positions.
+        # PaliGemma's image token can be OOV; PAD-mask those positions before
+        # the text embedding lookup (mirrors optimum-rbln's _preprocess_prefill).
         config = self.model.config
-        if is_multimodal is None:
-            is_multimodal = input_ids == config.image_token_id
-
         if config.image_token_id >= config.text_config.vocab_size:
-            llm_input_ids = input_ids.masked_fill(is_multimodal, PAD_TOKEN_ID)
-        else:
-            llm_input_ids = input_ids
-        inputs_embeds = self.model.get_input_embeddings()(llm_input_ids)
-
-        if multimodal_embeddings is None or len(multimodal_embeddings) == 0:
-            return inputs_embeds
-
-        # Flatten per-image embeddings into (num_mm_tokens, hidden_size).
-        mm_embeds = torch.cat(list(multimodal_embeddings)).to(
-            inputs_embeds.device, inputs_embeds.dtype
-        )
-        scatter_mask = is_multimodal.unsqueeze(-1).expand_as(inputs_embeds)
-        return inputs_embeds.masked_scatter(scatter_mask, mm_embeds)
+            input_ids = input_ids.masked_fill(is_multimodal, PAD_TOKEN_ID)
+        return self.model.get_input_embeddings()(input_ids)
 
     def _parse_and_validate_image_input(
         self, **kwargs: Any
